move page token functionalit yto pagination module, finish message pagination functionality
